git/task-clone-repo others, no need for ibmcloud region
The region parameter was causing problems
in an internal staging/test region.

in git/task-clone-repo, the ibmcloud login is only needed for the
ibmcloud iam oauth-tokens,
so no need for region or resource group to be configured.

in container-registry/configmap-check-registry,
doesn't need an ibmcloud -r region,
only the cr container registry region.

in container-registry/task-check-va-scan
only needs cr region, not ibmcloud region.

in toolchain/task-publish-deployable-mapping,
change to ibmcloud login --no-region